requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
certifi==2019.11.28       # via requests
chardet==3.0.4            # via requests
idna==2.8                 # via requests
imagehash==4.0
numpy==1.18.1             # via imagehash, opencv-python, pywavelets, scipy, torchvision
opencv-python==4.1.2.30
pillow==7.0.0             # via imagehash, torchvision
pywavelets==1.1.1         # via imagehash
requests==2.21.0
scipy==1.4.1              # via imagehash
six==1.14.0               # via imagehash, torchvision
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.0
urllib3==1.24.3           # via requests
